oAuth: long value has to be retrieved using getLong
Expiration times are Long in ms. Before this change, all tokens
are always considered expired when being loaded from the DB, due
to their expiration time being cut off.

+++ emailcommon/src/com/android/emailcommon/provider/Credential.java
@@ -90,7 +90,7 @@ public class Credential extends EmailContent implements Parcelable, BaseColumns 
        mProviderId = cursor.getString(CredentialQuery.PROVIDER_COLUMN_INDEX);
        mAccessToken = cursor.getString(CredentialQuery.ACCESS_TOKEN_COLUMN_INDEX);
        mRefreshToken = cursor.getString(CredentialQuery.REFRESH_TOKEN_COLUMN_INDEX);
-       mExpiration = cursor.getInt(CredentialQuery.EXPIRATION_COLUMN_INDEX);
+       mExpiration = cursor.getLong(CredentialQuery.EXPIRATION_COLUMN_INDEX);
    }
 
    /**
